moving sorting to the sources file
