+++ lister/srcs.go
@@ -1,5 +1,44 @@
 package lister
 
+import (
+	"cmp"
+	"math"
+	"slices"
+	"strings"
+)
+
+// In-place sorting of 'sources' based on given order of elements 'desiredOrder'
+//
+// # Omitted elements are placed after given elements
+//
+// # Duplicate elements use last-most given order
+//
+// Example:
+//
+//	sources := []string{"a", "b", "c", "x"}
+//	desiredOrder := []string{"b", "a", "c"}
+//	sortSources(sources, desiredOrder)
+//	// sources is now []string{"b", "a", "c", "x"}
+func sortSources(sources, desiredOrder []string) {
+	if desiredOrder == nil || len(desiredOrder) == 0 {
+		return
+	}
+	m := make(map[string]int)
+	for i, s := range desiredOrder {
+		m[strings.ToLower(s)] = i
+	}
+	getOrder := func(s string) int {
+		order, exists := m[strings.ToLower(s)]
+		if !exists {
+			return math.MaxInt
+		}
+		return order
+	}
+	slices.SortFunc(sources, func(a, b string) int {
+		return cmp.Compare(getOrder(a), getOrder(b))
+	})
+}
+
 func srcs(opts ListOptions) []string {
 	var srcs []string
 	count := 0
@@ -36,5 +75,8 @@ func srcs(opts ListOptions) []string {
 		srcs[i] = "zoxide"
 		i++
 	}
+	// TODO: if we have configured sorting, do so now
+	sortSources(srcs, []string{"tmuxinator", "tmux", "config", "zoxide"})
+
 	return srcs
 }
